# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ContactSettings {
  email: string;
  phone: string;
  whatsappNumber: string;
  location: string;
  businessHoursWeekdays: string;
  businessHoursWeekend: string;
}

export interface SocialLink {
  platform: string;
  label: string;
  url: string;
}

export interface PlanFeature {
  es: string;
  en: string;
}

export interface ServicePlan {
  slug: string;
  name: string;
  /** @nullable */
  priceCop: number | null;
  priceLabelEs: string;
  priceLabelEn: string;
  currency: string;
  /** @nullable */
  periodEs: string | null;
  /** @nullable */
  periodEn: string | null;
  /** @nullable */
  badgeEs: string | null;
  /** @nullable */
  badgeEn: string | null;
  isCustom: boolean;
  features: PlanFeature[];
}

export interface ServiceCategory {
  slug: string;
  nameEs: string;
  nameEn: string;
  plans: ServicePlan[];
}

export interface SiteContent {
  contact: ContactSettings;
  socialLinks: SocialLink[];
  services: ServiceCategory[];
}

export type ReviewLanguage = typeof ReviewLanguage[keyof typeof ReviewLanguage];


export const ReviewLanguage = {
  es: 'es',
  en: 'en',
} as const;

export interface Review {
  id: number;
  authorName: string;
  /** @nullable */
  company: string | null;
  quote: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  stars: number;
  language: ReviewLanguage;
  createdAt: string;
}

export interface ReviewsResponse {
  reviews: Review[];
  average: number;
  total: number;
}

export type CreateReviewRequestLanguage = typeof CreateReviewRequestLanguage[keyof typeof CreateReviewRequestLanguage];


export const CreateReviewRequestLanguage = {
  es: 'es',
  en: 'en',
} as const;

export interface CreateReviewRequest {
  /**
     * @minLength 2
     * @maxLength 120
     */
  authorName: string;
  /**
     * @maxLength 160
     * @nullable
     */
  company?: string | null;
  /**
     * @minLength 10
     * @maxLength 1200
     */
  quote: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  stars: number;
  language: CreateReviewRequestLanguage;
}

export type CreateContactMessageRequestLanguage = typeof CreateContactMessageRequestLanguage[keyof typeof CreateContactMessageRequestLanguage];


export const CreateContactMessageRequestLanguage = {
  es: 'es',
  en: 'en',
} as const;

export interface CreateContactMessageRequest {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /** @maxLength 254 */
  email: string;
  /**
     * @maxLength 40
     * @nullable
     */
  phone?: string | null;
  /**
     * @maxLength 160
     * @nullable
     */
  company?: string | null;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  message: string;
  language: CreateContactMessageRequestLanguage;
}

export interface SubmissionResponse {
  success: boolean;
  id: number;
  message: string;
}

export interface ErrorResponse {
  message: string;
}

/**
 * Invalid request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Internal server error
 */
export type ServerErrorResponse = ErrorResponse;

export type GetReviewsParams = {
language?: GetReviewsLanguage;
};

export type GetReviewsLanguage = typeof GetReviewsLanguage[keyof typeof GetReviewsLanguage];


export const GetReviewsLanguage = {
  es: 'es',
  en: 'en',
} as const;