# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: content
    description: Public site content
  - name: contact
    description: Contact messages
  - name: reviews
    description: Customer reviews
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /content:
    get:
      operationId: getSiteContent
      tags: [content]
      summary: Get public site content
      responses:
        "200":
          description: Contact information, social links, and service plans
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteContent"
        "500":
          $ref: "#/components/responses/ServerError"
  /reviews:
    get:
      operationId: getReviews
      tags: [reviews]
      summary: Get published reviews
      parameters:
        - name: language
          in: query
          required: false
          schema:
            type: string
            enum: [es, en]
      responses:
        "200":
          description: Published reviews and rating summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for moderation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewRequest"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /contact:
    post:
      operationId: createContactMessage
      tags: [contact]
      summary: Submit a contact message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactMessageRequest"
      responses:
        "201":
          description: Contact message saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
components:
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ContactSettings:
      type: object
      properties:
        email:
          type: string
        phone:
          type: string
        whatsappNumber:
          type: string
        location:
          type: string
        businessHoursWeekdays:
          type: string
        businessHoursWeekend:
          type: string
      required: [email, phone, whatsappNumber, location, businessHoursWeekdays, businessHoursWeekend]
    SocialLink:
      type: object
      properties:
        platform:
          type: string
        label:
          type: string
        url:
          type: string
      required: [platform, label, url]
    PlanFeature:
      type: object
      properties:
        es:
          type: string
        en:
          type: string
      required: [es, en]
    ServicePlan:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        priceCop:
          type: [number, "null"]
        priceLabelEs:
          type: string
        priceLabelEn:
          type: string
        currency:
          type: string
        periodEs:
          type: [string, "null"]
        periodEn:
          type: [string, "null"]
        badgeEs:
          type: [string, "null"]
        badgeEn:
          type: [string, "null"]
        isCustom:
          type: boolean
        features:
          type: array
          items:
            $ref: "#/components/schemas/PlanFeature"
      required: [slug, name, priceCop, priceLabelEs, priceLabelEn, currency, periodEs, periodEn, badgeEs, badgeEn, isCustom, features]
    ServiceCategory:
      type: object
      properties:
        slug:
          type: string
        nameEs:
          type: string
        nameEn:
          type: string
        plans:
          type: array
          items:
            $ref: "#/components/schemas/ServicePlan"
      required: [slug, nameEs, nameEn, plans]
    SiteContent:
      type: object
      properties:
        contact:
          $ref: "#/components/schemas/ContactSettings"
        socialLinks:
          type: array
          items:
            $ref: "#/components/schemas/SocialLink"
        services:
          type: array
          items:
            $ref: "#/components/schemas/ServiceCategory"
      required: [contact, socialLinks, services]
    Review:
      type: object
      properties:
        id:
          type: number
        authorName:
          type: string
        company:
          type: [string, "null"]
        quote:
          type: string
        stars:
          type: number
          minimum: 1
          maximum: 5
        language:
          type: string
          enum: [es, en]
        createdAt:
          type: string
      required: [id, authorName, company, quote, stars, language, createdAt]
    ReviewsResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        average:
          type: number
          format: float
        total:
          type: number
      required: [reviews, average, total]
    CreateReviewRequest:
      type: object
      properties:
        authorName:
          type: string
          minLength: 2
          maxLength: 120
        company:
          type: [string, "null"]
          maxLength: 160
        quote:
          type: string
          minLength: 10
          maxLength: 1200
        stars:
          type: number
          minimum: 1
          maximum: 5
        language:
          type: string
          enum: [es, en]
      required: [authorName, quote, stars, language]
    CreateContactMessageRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        email:
          type: string
          maxLength: 254
        phone:
          type: [string, "null"]
          maxLength: 40
        company:
          type: [string, "null"]
          maxLength: 160
        message:
          type: string
          minLength: 10
          maxLength: 2000
        language:
          type: string
          enum: [es, en]
      required: [name, email, message, language]
    SubmissionResponse:
      type: object
      properties:
        success:
          type: boolean
        id:
          type: number
        message:
          type: string
      required: [success, id, message]
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]