# lib/api-zod/src/generated/types/contactSettings.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactSettings {
  email: string;
  phone: string;
  whatsappNumber: string;
  location: string;
  businessHoursWeekdays: string;
  businessHoursWeekend: string;
}