# lib/api-zod/src/generated/types/createContactMessageRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateContactMessageRequestLanguage } from './createContactMessageRequestLanguage';

export interface CreateContactMessageRequest {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /** @maxLength 254 */
  email: string;
  /**
     * @maxLength 40
     * @nullable
     */
  phone?: string | null;
  /**
     * @maxLength 160
     * @nullable
     */
  company?: string | null;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  message: string;
  language: CreateContactMessageRequestLanguage;
}