# lib/api-zod/src/generated/types/createReviewRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateReviewRequestLanguage } from './createReviewRequestLanguage';

export interface CreateReviewRequest {
  /**
     * @minLength 2
     * @maxLength 120
     */
  authorName: string;
  /**
     * @maxLength 160
     * @nullable
     */
  company?: string | null;
  /**
     * @minLength 10
     * @maxLength 1200
     */
  quote: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  stars: number;
  language: CreateReviewRequestLanguage;
}