# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewLanguage } from './reviewLanguage';

export interface Review {
  id: number;
  authorName: string;
  /** @nullable */
  company: string | null;
  quote: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  stars: number;
  language: ReviewLanguage;
  createdAt: string;
}