# lib/api-zod/src/generated/types/reviewsResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Review } from './review';

export interface ReviewsResponse {
  reviews: Review[];
  average: number;
  total: number;
}