# lib/api-zod/src/generated/types/serviceCategory.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ServicePlan } from './servicePlan';

export interface ServiceCategory {
  slug: string;
  nameEs: string;
  nameEn: string;
  plans: ServicePlan[];
}