# lib/api-zod/src/generated/types/servicePlan.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlanFeature } from './planFeature';

export interface ServicePlan {
  slug: string;
  name: string;
  /** @nullable */
  priceCop: number | null;
  priceLabelEs: string;
  priceLabelEn: string;
  currency: string;
  /** @nullable */
  periodEs: string | null;
  /** @nullable */
  periodEn: string | null;
  /** @nullable */
  badgeEs: string | null;
  /** @nullable */
  badgeEn: string | null;
  isCustom: boolean;
  features: PlanFeature[];
}