# lib/api-zod/src/generated/types/siteContent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContactSettings } from './contactSettings';
import type { ServiceCategory } from './serviceCategory';
import type { SocialLink } from './socialLink';

export interface SiteContent {
  contact: ContactSettings;
  socialLinks: SocialLink[];
  services: ServiceCategory[];
}